requirements_backup.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-PyMongo==2.3.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
torch>=2.0.0
transformers>=4.28.1
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.0
faster-whisper>=0.10.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
easyocr>=1.7.1
opencv-python-headless>=4.10.0